# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Riviera Building Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /receipt-vouchers:
    get:
      operationId: listReceiptVouchers
      tags: [receiptVouchers]
      responses:
        "200":
          description: List of receipt vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReceiptVoucher"
    post:
      operationId: createReceiptVoucher
      tags: [receiptVouchers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptVoucherInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptVoucher"
  /receipt-vouchers/{id}:
    get:
      operationId: getReceiptVoucher
      tags: [receiptVouchers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt voucher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptVoucher"
        "404":
          description: Not found
    patch:
      operationId: updateReceiptVoucher
      tags: [receiptVouchers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiptVoucherUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptVoucher"
    delete:
      operationId: deleteReceiptVoucher
      tags: [receiptVouchers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Payment Vouchers ────────────────────────────────────────────────────────
  /payment-vouchers:
    get:
      operationId: listPaymentVouchers
      tags: [paymentVouchers]
      responses:
        "200":
          description: List of payment vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentVoucher"
    post:
      operationId: createPaymentVoucher
      tags: [paymentVouchers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVoucherInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVoucher"
  /payment-vouchers/{id}:
    get:
      operationId: getPaymentVoucher
      tags: [paymentVouchers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment voucher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVoucher"
        "404":
          description: Not found
    patch:
      operationId: updatePaymentVoucher
      tags: [paymentVouchers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVoucherUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVoucher"
    delete:
      operationId: deletePaymentVoucher
      tags: [paymentVouchers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Cash Fund ───────────────────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Audit Log ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ReceiptVoucher:
      type: object
      required: [id, voucherNumber, date, payerName, amount, currency, amountILS, paymentMethod]
      properties:
        id:
          type: integer
        voucherNumber:
          type: string
        date:
          type: string
        payerName:
          type: string
        tenantId:
          type: ["integer", "null"]
        tenantName:
          type: ["string", "null"]
        contractId:
          type: ["integer", "null"]
        contractNumber:
          type: ["string", "null"]
        unitNumber:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        amountILS:
          type: number
        paymentMethod:
          type: string
          enum: [cash, bank_transfer, cheque]
        chequeNumber:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        chequeDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        accountHolderName:
          type: ["string", "null"]
        previousBalance:
          type: ["number", "null"]
        newBalance:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ReceiptVoucherInput:
      type: object
      required: [date, payerName, amount, currency, exchangeRate, amountILS, paymentMethod]
      properties:
        date:
          type: string
        payerName:
          type: string
        tenantId:
          type: ["integer", "null"]
        contractId:
          type: ["integer", "null"]
        amount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        amountILS:
          type: number
        paymentMethod:
          type: string
          enum: [cash, bank_transfer, cheque]
        chequeNumber:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        chequeDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        accountHolderName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ReceiptVoucherUpdate:
      type: object
      properties:
        date:
          type: string
        payerName:
          type: string
        amount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        amountILS:
          type: number
        paymentMethod:
          type: string
          enum: [cash, bank_transfer, cheque]
        chequeNumber:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        chequeDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        accountHolderName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ─── Payment Vouchers ──────────────────────────────────────────────────────
    PaymentVoucher:
      type: object
      required: [id, voucherNumber, date, beneficiaryName, amount, currency, amountILS, paymentMethod, category]
      properties:
        id:
          type: integer
        voucherNumber:
          type: string
        date:
          type: string
        beneficiaryName:
          type: string
        amount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        amountILS:
          type: number
        paymentMethod:
          type: string
          enum: [cash, bank_transfer, cheque]
        category:
          type: string
        chequeNumber:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        chequeDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentVoucherInput:
      type: object
      required: [date, beneficiaryName, amount, currency, exchangeRate, amountILS, paymentMethod, category]
      properties:
        date:
          type: string
        beneficiaryName:
          type: string
        amount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        amountILS:
          type: number
        paymentMethod:
          type: string
          enum: [cash, bank_transfer, cheque]
        category:
          type: string
        chequeNumber:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        chequeDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    PaymentVoucherUpdate:
      type: object
      properties:
        date:
          type: string
        beneficiaryName:
          type: string
        amount:
          type: number
        currency:
          type: string
          enum: [ILS, USD, JOD]
        exchangeRate:
          type: number
        amountILS:
          type: number
        paymentMethod:
          type: string
          enum: [cash, bank_transfer, cheque]
        category:
          type: string
        chequeNumber:
          type: ["string", "null"]
        bankName:
          type: ["string", "null"]
        chequeDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ─── Cash Fund ─────────────────────────────────────────────────────────────
...
    Document:
      type: object
      required: [id, name, entityType, entityId, fileType]
      properties:
        id:
          type: integer
        name:
          type: string
        entityType:
          type: string
          enum: [tenant, contract, unit, general]
        entityId:
          type: ["integer", "null"]
        entityName:
          type: ["string", "null"]
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        fileUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    DocumentInput:
      type: object
      required: [name, entityType, fileType]
      properties:
        name:
          type: string
        entityType:
          type: string
          enum: [tenant, contract, unit, general]
        entityId:
          type: ["integer", "null"]
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        fileUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ─── Audit Log ─────────────────────────────────────────────────────────────
...